pyproject.toml
[project]
name = "new-blog-rag-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.116.0",
    "openai>=1.93.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
